+++ OpenGL 3D Renderer - Ubay/MyClasses/MyForceGenerator/MyGravityGenerator/MyGravityGenerator.cpp
@@ -10,7 +10,7 @@ MyGravityGenerator::MyGravityGenerator(MyVector3 gravity) : gravity(gravity) {}
 //* ╔═════════╗
 //* ║ Methods ║
 //* ╚═════════╝
-void MyGravityGenerator::updateForce(MyParticle* targetParticle, float time) {
+void MyGravityGenerator::updateForce(MyParticle* targetParticle, double time) {
     if (targetParticle->getMass() <= 0) return;
     targetParticle->addForce(this->gravity * targetParticle->getMass());
 }
