+++ OpenGL 3D Renderer - Ubay/MyClasses/MyForceGenerator/MyGravityGenerator/MyGravityGenerator.hpp
@@ -24,7 +24,7 @@ public:
     //* ║ Methods ║
     //* ╚═════════╝
 public:
-    void updateForce(MyParticle* targetParticle, float time) override;
+    void updateForce(MyParticle* targetParticle, double time) override;
 
 private:
     //* ╔═══════════════════╗
